Add default cities and categories in DataSeedingListtener

+++ src/main/java/com/example/case_module4/controller/BookingRestController.java
@@ -10,6 +10,8 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.swing.text.html.Option;
+import java.awt.print.Book;
 import java.time.LocalDate;
 import java.util.Optional;
 
@@ -58,6 +60,13 @@ public class BookingRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Booking> removeBooking(@PathVariable Long id) {
+        Optional<Booking> bookingOptional = bookingService.findById(id);
+        if (!bookingOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        Booking booking = bookingOptional.get();
+        booking.getRoom().setIsAvailable(
+                !LocalDate.now().isAfter(booking.getCheckIn()) || !LocalDate.now().isBefore(booking.getCheckOut()));
         bookingService.deleteById(id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
